+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 ### Added
 - New `tag.select.merge` function to efficiently merge sorted feature streams (see #68).
+- New `tag.locus.loci` function to efficiently determine locus coordinates from multiple annotation streams (see #71).
 
 ### Changed
 - Minor updates to compensate for a couple years' worth of neglect (see #64).